+++ packages/xrpl/src/Wallet/signer.ts
@@ -7,7 +7,6 @@ import { verify } from 'ripple-keypairs'
 import { ValidationError } from '../errors'
 import { Signer } from '../models/common'
 import { Transaction, validate } from '../models/transactions'
-import { Batch, BatchSigner, validateBatch } from '../models/transactions/batch'
 
 /**
  * Takes several transactions with Signer fields (in object or blob form) and creates a
@@ -54,57 +53,6 @@ function multisign(transactions: Array<Transaction | string>): string {
   validateTransactionEquivalence(decodedTransactions)
 
   return encode(getTransactionWithAllSigners(decodedTransactions))
-}
-
-/**
- * Takes several transactions with BatchSigners fields (in object or blob form) and creates a
- * single transaction with all BatchSigners that then gets signed and returned.
- *
- * @param transactions The transactions to combine `BatchSigners` values on.
- * @returns A single signed Transaction which has all BatchSigners from transactions within it.
- * @throws ValidationError if:
- * - There were no transactions given to sign
- * @category Signing
- */
-function combineBatchSignatures(transactions: Array<Batch | string>): string {
-  if (transactions.length === 0) {
-    throw new ValidationError('There were 0 transactions to combine')
-  }
-
-  const decodedTransactions: Transaction[] = transactions.map(
-    (txOrBlob: string | Transaction) => {
-      return getDecodedTransaction(txOrBlob)
-    },
-  )
-
-  decodedTransactions.forEach((tx) => {
-    if (tx.TransactionType !== 'Batch') {
-      throw new ValidationError('TransactionType must be `Batch`.')
-    }
-    /*
-     * This will throw a more clear error for JS users if any of the supplied transactions has incorrect formatting
-     */
-    // eslint-disable-next-line @typescript-eslint/consistent-type-assertions -- validate does not accept Transaction type
-    validateBatch(tx as unknown as Record<string, unknown>)
-    if (tx.Signers == null || tx.Signers.length === 0) {
-      throw new ValidationError(
-        "For multisigning all transactions must include a Signers field containing an array of signatures. You may have forgotten to pass the 'forMultisign' parameter when signing.",
-      )
-    }
-
-    if (tx.SigningPubKey !== '') {
-      throw new ValidationError(
-        'SigningPubKey must be an empty string for all transactions when multisigning.',
-      )
-    }
-  })
-
-  // eslint-disable-next-line @typescript-eslint/consistent-type-assertions -- checked above
-  const batchTransactions = decodedTransactions as Batch[]
-
-  validateBatchTransactionEquivalence(batchTransactions)
-
-  return encode(getTransactionWithAllBatchSigners(batchTransactions))
 }
 
 /**
@@ -167,32 +115,6 @@ function validateTransactionEquivalence(transactions: Transaction[]): void {
   }
 }
 
-/**
- * The transactions should all be equal except for the 'Signers' field.
- *
- * @param transactions - An array of Transactions which are expected to be equal other than 'Signers'.
- * @throws ValidationError if the transactions are not equal in any field other than 'Signers'.
- */
-function validateBatchTransactionEquivalence(transactions: Batch[]): void {
-  const exampleTransaction = JSON.stringify({
-    Flags: transactions[0].Flags,
-    TxIDs: transactions[0].TxIDs,
-  })
-  if (
-    transactions
-      .slice(1)
-      .some(
-        (tx) =>
-          JSON.stringify({ Flags: tx.Flags, TxIDs: tx.TxIDs }) !==
-          exampleTransaction,
-      )
-  ) {
-    throw new ValidationError(
-      'Flags and TxIDs is not the same for all provided transactions',
-    )
-  }
-}
-
 function getTransactionWithAllSigners(
   transactions: Transaction[],
 ): Transaction {
@@ -202,15 +124,6 @@ function getTransactionWithAllSigners(
     .sort(compareSigners)
 
   return { ...transactions[0], Signers: sortedSigners }
-}
-
-function getTransactionWithAllBatchSigners(transactions: Batch[]): Transaction {
-  // Signers must be sorted in the combined transaction - See compareSigners' documentation for more details
-  const sortedSigners: BatchSigner[] = transactions
-    .flatMap((tx) => tx.BatchSigners ?? [])
-    .sort(compareBatchSigners)
-
-  return { ...transactions[0], BatchSigners: sortedSigners }
 }
 
 /**
@@ -226,22 +139,6 @@ function getTransactionWithAllBatchSigners(transactions: Batch[]): Transaction {
 function compareSigners(left: Signer, right: Signer): number {
   return addressToBigNumber(left.Signer.Account).comparedTo(
     addressToBigNumber(right.Signer.Account),
-  )
-}
-
-/**
- * If presented in binary form, the BatchSigners array must be sorted based on
- * the numeric value of the signer addresses, with the lowest value first.
- * (If submitted as JSON, the submit_multisigned method handles this automatically.)
- * https://xrpl.org/multi-signing.html.
- *
- * @param left - A BatchSigner to compare with.
- * @param right - A second BatchSigner to compare with.
- * @returns 1 if left \> right, 0 if left = right, -1 if left \< right, and null if left or right are NaN.
- */
-function compareBatchSigners(left: BatchSigner, right: BatchSigner): number {
-  return addressToBigNumber(left.BatchSigner.Account).comparedTo(
-    addressToBigNumber(right.BatchSigner.Account),
   )
 }
 
@@ -263,4 +160,4 @@ function getDecodedTransaction(txOrBlob: Transaction | string): Transaction {
   return decode(txOrBlob) as unknown as Transaction
 }
 
-export { combineBatchSignatures, verifySignature, multisign }
+export { verifySignature, multisign }
